netfilter: Create new type for kernel pointers.
In the netfilter subsystem there are many structs inside the UAPI
headers which contain kernel pointers. This causes misalignment between
userspace and kernel space of these structures when using the PCuABI,
due to the differing sizes of the pointer in kernel space, and the
capability in userspace. This leads to issues with some netlink messages
sent between user/kernel space.

In order to fix this, we require a new type which is the same size from
the perspective of both the user and the kernel, and is able to hold a kernel
pointer regardless of the ABI which is being used (non-purecap, hybrid
pc-userspace/non-pc kernel, or a fully purecap kernel and userpsace).

In order to achieve this we create a new type which is the size of a capability
when using CONFIG_CHERI_PURECAP_UABI, and the size of a pointer
if not. We will then replace the kernel pointers in the UAPI structs with
this new type. Doing this will maintain consistent struct sizing, even
if a fully purecap kernel is used.

The drawback to this approach is that this method is unnecessary in the case
where we have a purecap userspace and non-purecap kernel, since the kernel
pointer in the structs are never actually touched by userspace, and so are not
required to be the size of a capability. However, it is necessary to guarantee
that the UAPI remains stable no matter what the kernel ABI uses.

The use of x_tables.h and netfilter.h appears to be mutually
exclusive within the kernel.This is why it is required to add the new
definition in both netfilter.h and x_tables.h. Since many conflicts between
definitions exist when including netfilter.h in x_tables.h. This overlap is
presumably is an artefact of xtables being superseded by nftables.

Signed-off-by: Joshua Lant <joshualant@gmail.com>