xt plugins: pointers to unions in uapi struct
On morello architecture, use of kernel pointers in the uapi
structures is not permitted, due to different alignment requirements
between kernel and userspace. These pointers should be the new type
__nf_kptr_t, whose size changes depening on the pc/non-pcUABI.
Modify these to use a union which will be accessed in the kernel using
the original member pointer, but will actually be of size __nf_kptr_t,
avoiding heavy casting needed when using to new type directly.

In order to have the __nf_kptr_t type included, we need to move
x_tables.h include into the headers rather than the c files.

Signed-off-by: Joshua Lant <joshualant@gmail.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>