nvme: Silence warning in !PCuABI
Commit "io_uring: Add explicit capability checks" modified
io_uring_cmd_import_fixed() to take a user pointer instead of an
unsigned long. The NVMe driver will need to be adapted accordingly
(as per the comment) to support PCuABI, but that change also
introduced a warning in !PCuABI. Silence it with an appropriate
cast.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>