aio: Silence integer conversion warning
Commit "aio: Change aio_context_t to a user pointer" ensured that
the compat io_setup handler always manipulates the context as
compat_aio_context_t. However, that type is 32-bit in compat32 and
printing it with %lu will trigger a warning. Add a cast to unsigned
long to avoid that warning.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>